--- dnxjay/code/conversions/gps_distance.py
@@ -1,16 +1,0 @@
-import math
-
-def haversine(lat1, lon1, lat2, lon2):
-    """Calculate the Haversine distance between two GPS points in meters."""
-    radius = 6371e3  # Earth's radius in meters
-
-    phi1 = math.radians(lat1)
-    phi2 = math.radians(lat2)
-    delta_phi = math.radians(lat2 - lat1)
-    delta_lambda = math.radians(lon2 - lon1)
-
-    a = math.sin(delta_phi / 2) ** 2 + math.cos(phi1) * math.cos(phi2) * math.sin(delta_lambda / 2) ** 2
-    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))
-
-    distance = radius * c
-    return distance
