--- dnxjay/code/conversions/unit_converter.py
@@ -1,7 +1,0 @@
-def feet_to_meters(feet):
-    """Convert feet to meters."""
-    return feet * 0.3048
-
-def yards_to_meters(yards):
-    """Convert yards to meters."""
-    return yards * 0.9144
