+++ dnxjay/code/python_modules/gps_distance.py
@@ -1,0 +1,27 @@
+# gps_distance.py
+import math
+
+def haversine(lat1, lon1, lat2, lon2):
+    """Calculate the Haversine distance between two GPS points in meters.
+    
+    Args:
+        lat1 (float): Latitude of the first point in decimal degrees.
+        lon1 (float): Longitude of the first point in decimal degrees.
+        lat2 (float): Latitude of the second point in decimal degrees.
+        lon2 (float): Longitude of the second point in decimal degrees.
+        
+    Returns:
+        float: Distance between the two points in meters.
+    """
+    radius = 6371e3  # Earth's radius in meters
+
+    phi1 = math.radians(lat1)
+    phi2 = math.radians(lat2)
+    delta_phi = math.radians(lat2 - lat1)
+    delta_lambda = math.radians(lon2 - lon1)
+
+    a = math.sin(delta_phi / 2) ** 2 + math.cos(phi1) * math.cos(phi2) * math.sin(delta_lambda / 2) ** 2
+    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1 - a))
+
+    distance = radius * c
+    return distance
