+++ dnxjay/code/python_modules/unit_converter.py
@@ -1,0 +1,23 @@
+# unit_converter.py
+
+def feet_to_meters(feet):
+    """Convert feet to meters.
+    
+    Args:
+        feet (float): Distance in feet.
+        
+    Returns:
+        float: Distance in meters.
+    """
+    return feet * 0.3048
+
+def yards_to_meters(yards):
+    """Convert yards to meters.
+    
+    Args:
+        yards (float): Distance in yards.
+        
+    Returns:
+        float: Distance in meters.
+    """
+    return yards * 0.9144
